feat(docs): add comprehensive Highlight component examples and documentation
- Add case-insensitive and case-sensitive matching examples
- Add variant-specific examples for highlight and emphasis styles
- Add transform function example for accent-insensitive search
- Update main documentation with improved structure and clearer explanations
- Reorganize examples from basic usage to advanced features
- Update props table with better descriptions and type information
