--- packages/react-docs/pages/components/highlight/basic.js
@@ -1,15 +1,0 @@
-import { Highlight, Stack } from '@tonic-ui/react';
-import React from 'react';
-
-const App = () => (
-  <Stack direction="column" spacing="4x">
-    <Highlight query="highlight">
-      This text contains the word highlight which will be highlighted.
-    </Highlight>
-    <Highlight query="highlight">
-      The &ldquo;Highlight&rdquo; component is useful for search results and text matching.
-    </Highlight>
-  </Stack>
-);
-
-export default App;
