+++ packages/react-docs/pages/components/highlight/case-insensitive-matching.js
@@ -1,0 +1,10 @@
+import { Highlight } from '@tonic-ui/react';
+import React from 'react';
+
+const App = () => (
+  <Highlight query="react">
+    React, REACT, react, and ReAcT will all be highlighted because the search is case-insensitive by default.
+  </Highlight>
+);
+
+export default App;
