+++ packages/react-docs/pages/components/highlight/query.js
@@ -1,0 +1,10 @@
+import { Highlight } from '@tonic-ui/react';
+import React from 'react';
+
+const App = () => (
+  <Highlight query="highlight">
+    This text contains the word highlight which will be highlighted.
+  </Highlight>
+);
+
+export default App;
