+++ packages/react-docs/pages/components/highlight/variant-emphasis.js
@@ -1,0 +1,10 @@
+import { Highlight } from '@tonic-ui/react';
+import React from 'react';
+
+const App = () => (
+  <Highlight variant="emphasis" query="emphasis">
+    The emphasis variant is intended to highlight text, such as in auto-complete results.
+  </Highlight>
+);
+
+export default App;
