+++ packages/react-docs/pages/components/highlight/variant-highlight.js
@@ -1,0 +1,10 @@
+import { Highlight } from '@tonic-ui/react';
+import React from 'react';
+
+const App = () => (
+  <Highlight variant="highlight" query="highlight">
+    The highlight variant marks search keywords with a yellow background and is typically used for table search results.
+  </Highlight>
+);
+
+export default App;
